+++ src/Console/Commands/ManagementCaronte.php
@@ -22,6 +22,9 @@ class ManagementCaronte extends Command
                 'Update user',
                 'Delete user',
                 'Sync user roles',
+                'List suite roles',
+                'List suite users',
+                'Sync suite user roles',
                 'Exit',
             ]);
 
@@ -36,6 +39,9 @@ class ManagementCaronte extends Command
                 'Update user' => $this->call('caronte:users:update'),
                 'Delete user' => $this->call('caronte:users:delete'),
                 'Sync user roles' => $this->call('caronte:users:roles:sync'),
+                'List suite roles' => $this->call('caronte:groups:roles:list'),
+                'List suite users' => $this->call('caronte:groups:users:list'),
+                'Sync suite user roles' => $this->call('caronte:groups:users:roles:sync'),
                 'Exit' => null,
             };