+++ tests/Feature/RouteRegistrationTest.php
@@ -23,5 +23,6 @@ class RouteRegistrationTest extends TestCase
         $this->assertTrue(Route::has('caronte.management.users.delete'));
         $this->assertTrue(Route::has('caronte.management.users.delete.direct'));
         $this->assertTrue(Route::has('caronte.management.users.roles.list'));
+        $this->assertTrue(Route::has('caronte.management.suite.users.applications.roles.sync'));
     }
 }